+++ pumpkin-world/src/chunk/io/chunk_file_manager.rs
@@ -1,6 +1,5 @@
 use std::{
     collections::BTreeMap,
-    error,
     io::ErrorKind,
     path::{Path, PathBuf},
     sync::Arc,
@@ -11,86 +10,17 @@ use log::{error, trace};
 use pumpkin_util::math::vector2::Vector2;
 use tokio::{
     fs::OpenOptions,
-    io::AsyncReadExt,
-    sync::{OnceCell, mpsc},
+    io::{AsyncReadExt, AsyncWriteExt},
+    sync::{OnceCell, RwLock, mpsc},
     task::JoinSet,
 };
-use tokio::{io::AsyncWriteExt, sync::RwLock};
 
 use crate::{
     chunk::{ChunkReadingError, ChunkWritingError},
     level::LevelFolder,
 };
 
-pub enum LoadedData<D, Err: error::Error>
-where
-    D: Send,
-{
-    /// The chunk data was loaded successfully
-    Loaded(D),
-    /// The chunk data was not found
-    Missing(Vector2<i32>),
-
-    /// An error occurred while loading the chunk data
-    Error((Vector2<i32>, Err)),
-}
-
-/// Trait to handle the IO of chunks
-/// for loading and saving chunks data
-/// can be implemented for different types of IO
-/// or with different optimizations
-///
-/// The `R` type is the type of the data that will be loaded/saved
-/// like ChunkData or EntityData
-#[async_trait]
-pub trait ChunkIO<D>
-where
-    Self: Send + Sync,
-    D: Send + Sized,
-{
-    /// Load the chunks data
-    async fn stream_chunks(
-        &self,
-        folder: &LevelFolder,
-        chunk_coords: &[Vector2<i32>],
-        channel: mpsc::Sender<LoadedData<D, ChunkReadingError>>,
-    );
-
-    /// Persist the chunks data
-    async fn save_chunks(
-        &self,
-        folder: &LevelFolder,
-        chunks_data: Vec<(Vector2<i32>, Arc<RwLock<D>>)>,
-    ) -> Result<(), ChunkWritingError>;
-
-    async fn clean_up_log(&self);
-
-    async fn await_tasks(&self);
-}
-
-/// Trait to serialize and deserialize the chunk data to and from bytes.
-///
-/// The `Data` type is the type of the data that will be updated or serialized/deserialized
-/// like ChunkData or EntityData
-
-#[async_trait]
-pub trait ChunkSerializer: Send + Sync + Sized + Default {
-    type Data: Send;
-
-    fn get_chunk_key(chunk: Vector2<i32>) -> String;
-
-    fn to_bytes(&self) -> Box<[u8]>;
-
-    fn from_bytes(bytes: &[u8]) -> Result<Self, ChunkReadingError>;
-
-    fn add_chunk_data(&mut self, chunk_data: &[Self::Data]) -> Result<(), ChunkWritingError>;
-
-    async fn stream_chunk_data(
-        &self,
-        chunks: &[Vector2<i32>],
-        channel: mpsc::Sender<LoadedData<Self::Data, ChunkReadingError>>,
-    );
-}
+use super::{ChunkIO, ChunkSerializer, LoadedData};
 
 /// A simple implementation of the ChunkSerializer trait
 /// that load and save the data from a file in the disk
