+++ pumpkin-world/src/lib.rs
@@ -4,7 +4,6 @@ use pumpkin_util::math::vector2::Vector2;
 pub mod biome;
 pub mod block;
 pub mod chunk;
-pub mod chunks_io;
 pub mod coordinates;
 pub mod cylindrical_chunk_iterator;
 pub mod dimension;
